refactor(iota-genesis-builder, iota-types): move stardust types from iota-genesis-builder to iota-types Separating snapshot file into outputHeader and snapshot. Rid of the BasicOutput dependency on output_header thought new_object_id method in output_header Merging timelock.rs into timelock module
