+++ crates/iota-genesis-builder/Cargo.toml
@@ -12,7 +12,6 @@ bcs.workspace = true
 camino.workspace = true
 clap.workspace = true
 fastcrypto.workspace = true
-hex.workspace = true
 itertools.workspace = true
 move-binary-format.workspace = true
 move-compiler.workspace = true
@@ -35,8 +34,8 @@ bigdecimal = "0.4.3"
 brotli = "6.0.0"
 fs_extra = "1.3.0"
 iota-sdk = { version = "1.1.4", default-features = false, features = ["irc_27", "irc_30", "std"] }
-num-rational = "0.4"
 packable = { version = "0.8.3", default-features = false, features = ["io"] }
+
 rand_pcg = "0.3.1"
 rand_seeder = "0.2.3"
 tracing-subscriber = "0.3.11"
