+++ crates/iota-genesis-builder/src/stardust/migration/tests/foundry.rs
@@ -16,6 +16,7 @@ use iota_types::{
     coin_manager::CoinManager,
     gas_coin::GAS,
     object::Object,
+    stardust::{stardust_to_iota_address, stardust_to_iota_address_owner},
     smr_coin::{SmrCoin, SMR},
 };
 use move_core_types::language_storage::TypeTag;
@@ -25,7 +26,7 @@ use crate::stardust::{
         tests::{create_foundry, run_migration},
         CoinType,
     },
-    types::{snapshot::OutputHeader, stardust_to_iota_address, stardust_to_iota_address_owner},
+    types::output_header::OutputHeader,
 };
 
 type PackageObject = Object;
