+++ crates/iota-genesis-builder/src/stardust/migration/tests/nft.rs
@@ -31,6 +31,14 @@ use iota_types::{
     dynamic_field::{derive_dynamic_field_id, DynamicFieldInfo},
     id::UID,
     object::{Object, Owner},
+    stardust::{
+        output::{
+            FixedPoint32, Irc27Metadata, Nft, NftOutput, ALIAS_OUTPUT_MODULE_NAME,
+            NFT_DYNAMIC_OBJECT_FIELD_KEY, NFT_DYNAMIC_OBJECT_FIELD_KEY_TYPE,
+            NFT_OUTPUT_MODULE_NAME,
+        },
+        stardust_to_iota_address,
+    },
     TypeTag,
 };
 use move_core_types::ident_str;
@@ -44,11 +52,7 @@ use crate::stardust::{
         },
         CoinType,
     },
-    types::{
-        snapshot::OutputHeader, stardust_to_iota_address, FixedPoint32, Irc27Metadata, Nft,
-        NftOutput, ALIAS_OUTPUT_MODULE_NAME, NFT_DYNAMIC_OBJECT_FIELD_KEY,
-        NFT_DYNAMIC_OBJECT_FIELD_KEY_TYPE, NFT_OUTPUT_MODULE_NAME,
-    },
+    types::output_header::OutputHeader,
 };
 
 fn migrate_nft(
@@ -360,7 +364,7 @@ fn nft_migration_with_valid_irc27_metadata() {
 
     let immutable_metadata = nft.immutable_metadata;
     assert_eq!(&immutable_metadata.media_type, metadata.media_type());
-    assert_eq!(immutable_metadata.uri.url, metadata.uri().to_string());
+    assert_eq!(immutable_metadata.uri.url(), metadata.uri().to_string());
     assert_eq!(&immutable_metadata.name, metadata.name());
     assert_eq!(&immutable_metadata.issuer_name, metadata.issuer_name());
     assert_eq!(
