+++ crates/iota-genesis-builder/src/stardust/migration/verification/alias.rs
@@ -11,22 +11,20 @@ use iota_types::{
     dynamic_field::{derive_dynamic_field_id, DynamicFieldInfo, Field},
     in_memory_storage::InMemoryStorage,
     object::Owner,
+    stardust::output::{
+        Alias, AliasOutput, ALIAS_DYNAMIC_OBJECT_FIELD_KEY, ALIAS_DYNAMIC_OBJECT_FIELD_KEY_TYPE,
+    },
     TypeTag,
 };
 
-use crate::stardust::{
-    migration::{
-        executor::FoundryLedgerData,
-        verification::{
-            created_objects::CreatedObjects,
-            util::{
-                verify_address_owner, verify_issuer_feature, verify_metadata_feature,
-                verify_native_tokens, verify_parent, verify_sender_feature,
-            },
+use crate::stardust::migration::{
+    executor::FoundryLedgerData,
+    verification::{
+        created_objects::CreatedObjects,
+        util::{
+            verify_address_owner, verify_issuer_feature, verify_metadata_feature,
+            verify_native_tokens, verify_parent, verify_sender_feature,
         },
-    },
-    types::{
-        Alias, AliasOutput, ALIAS_DYNAMIC_OBJECT_FIELD_KEY, ALIAS_DYNAMIC_OBJECT_FIELD_KEY_TYPE,
     },
 };
