+++ crates/iota-genesis-builder/src/stardust/mod.rs
@@ -5,7 +5,6 @@
 //! parsing Stardust UTXOs from a full-snapshot file, and converting
 //! them to the appropriate genesis objects.
 
-pub mod error;
 pub mod migration;
 pub mod native_token;
 pub mod parse;
