+++ crates/iota-genesis-builder/src/stardust/native_token/package_builder.rs
@@ -12,9 +12,10 @@ use std::{
 use anyhow::Result;
 use fs_extra::dir::{copy, CopyOptions};
 use iota_move_build::{BuildConfig, CompiledPackage, IotaPackageHooks};
+use iota_types::stardust::error::StardustError;
 use tempfile::tempdir;
 
-use crate::stardust::{error::StardustError, native_token::package_data::NativeTokenPackageData};
+use crate::stardust::native_token::package_data::NativeTokenPackageData;
 
 /// Builds and compiles a Stardust native token package.
 pub fn build_and_compile(package: NativeTokenPackageData) -> Result<CompiledPackage> {
