+++ crates/iota-genesis-builder/src/stardust/types/mod.rs
@@ -1,16 +1,6 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod address;
-pub mod alias;
-pub mod coin_kind;
-pub mod foundry;
-pub mod nft;
-pub mod output;
+pub mod output_header;
 pub mod snapshot;
-pub mod timelock;
 pub mod token_scheme;
-
-pub use address::*;
-pub use alias::*;
-pub use nft::*;
