+++ crates/iota-types/Cargo.toml
@@ -7,6 +7,9 @@ license = "Apache-2.0"
 publish = false
 
 [dependencies]
+iota-stardust-sdk = { version = "1.1.4", default-features = false, features = ["irc_27", "irc_30", "std"], package = "iota-sdk" }
+num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
+packable = { version = "0.8.3", default-features = false, features = ["io"] }
 anemo.workspace = true
 anyhow.workspace = true
 bcs.workspace = true
@@ -16,6 +19,7 @@ consensus-config.workspace = true
 derivative.workspace = true
 enum_dispatch.workspace = true
 eyre.workspace = true
+hex.workspace = true
 im.workspace = true
 indexmap.workspace = true
 itertools.workspace = true
@@ -29,7 +33,7 @@ move-vm-profiler.workspace = true
 move-vm-test-utils.workspace = true
 move-vm-types.workspace = true
 nonempty.workspace = true
-num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
+num-rational = "0.4"
 num-traits = "0.2.18"
 once_cell.workspace = true
 prometheus.workspace = true
