+++ crates/iota-types/src/lib.rs
@@ -78,6 +78,7 @@ pub mod programmable_transaction_builder;
 pub mod quorum_driver_types;
 pub mod randomness_state;
 pub mod signature;
+pub mod stardust;
 pub mod smr_coin;
 pub mod storage;
 pub mod timelock;
