+++ crates/iota-types/src/stardust/coin_kind.rs
@@ -16,14 +16,13 @@
 
 use std::mem::size_of;
 
-use iota_types::{
+use crate::{
     balance::Balance,
     gas_coin::GAS,
     object::{Object, ID_END_INDEX},
+    stardust::output::{AliasOutput, BasicOutput, NftOutput},
     timelock::timelock::TimeLock,
 };
-
-use crate::stardust::types::{output::BasicOutput, AliasOutput, NftOutput};
 
 /// Infer whether the object is a kind of gas coin.
 pub fn is_gas_coin_kind(object: &Object) -> bool {
@@ -49,19 +48,20 @@ pub fn get_gas_balance_maybe(object: &Object) -> Option<Balance> {
 }
 
 #[cfg(test)]
+#[cfg(feature = "test-utils")]
 mod tests {
 
     use iota_protocol_config::ProtocolConfig;
-    use iota_types::{
+
+    use super::*;
+    use crate::{
         base_types::{IotaAddress, ObjectID, TxContext},
         id::UID,
         object::Owner,
+        stardust::timelock::to_genesis_object,
     };
 
-    use super::*;
-    use crate::stardust::{migration::CoinType, types::timelock::to_genesis_object};
-
-    fn nft_output(balance: u64, coin_type: CoinType) -> anyhow::Result<Object> {
+    fn nft_output(balance: u64) -> anyhow::Result<Object> {
         let id = UID::new(ObjectID::random());
         let balance = Balance::new(balance);
         let output = NftOutput {
@@ -77,25 +77,25 @@ mod tests {
             &ProtocolConfig::get_for_min_version(),
             &TxContext::random_for_testing_only(),
             1.into(),
-            coin_type,
+            GAS::type_tag(),
         )
     }
 
     #[test]
     fn is_coin_kind_nft_output() {
-        let object = nft_output(100, CoinType::Iota).unwrap();
+        let object = nft_output(100).unwrap();
         assert!(is_gas_coin_kind(&object));
     }
 
     #[test]
     fn get_gas_balance_nft_output() {
         let value = 100;
-        let object = nft_output(value, CoinType::Iota).unwrap();
+        let object = nft_output(value).unwrap();
         let gas_coin_balance = get_gas_balance_maybe(&object).unwrap();
         assert_eq!(gas_coin_balance.value(), value);
     }
 
-    fn alias_output(balance: u64, coin_type: CoinType) -> anyhow::Result<Object> {
+    fn alias_output(balance: u64) -> anyhow::Result<Object> {
         let id = UID::new(ObjectID::random());
         let balance = Balance::new(balance);
         let output = AliasOutput {
@@ -106,27 +106,27 @@ mod tests {
         output.to_genesis_object(
             Owner::AddressOwner(IotaAddress::ZERO),
             &ProtocolConfig::get_for_min_version(),
-            &TxContext::random_for_testSing_only(),
+            &TxContext::random_for_testing_only(),
             1.into(),
-            &coin_type,
+            GAS::type_tag(),
         )
     }
 
     #[test]
     fn is_coin_kind_alias_output() {
-        let object = alias_output(100, CoinType::Iota).unwrap();
+        let object = alias_output(100).unwrap();
         assert!(is_gas_coin_kind(&object));
     }
 
     #[test]
     fn get_gas_balance_alias_output() {
         let value = 100;
-        let object = alias_output(value, CoinType::Iota).unwrap();
+        let object = alias_output(value).unwrap();
         let gas_coin_balance = get_gas_balance_maybe(&object).unwrap();
         assert_eq!(gas_coin_balance.value(), value);
     }
 
-    fn basic_output(balance: u64, coin_type: CoinType) -> anyhow::Result<Object> {
+    fn basic_output(balance: u64) -> anyhow::Result<Object> {
         let id = UID::new(ObjectID::random());
         let balance = Balance::new(balance);
         let output = BasicOutput {
@@ -145,20 +145,20 @@ mod tests {
             &ProtocolConfig::get_for_min_version(),
             &TxContext::random_for_testing_only(),
             1.into(),
-            &coin_type,
+            GAS::type_tag(),
         )
     }
 
     #[test]
     fn is_coin_kind_basic_output() {
-        let object = basic_output(100, CoinType::Iota).unwrap();
+        let object = basic_output(100).unwrap();
         assert!(is_gas_coin_kind(&object));
     }
 
     #[test]
     fn get_gas_balance_basic_output() {
         let value = 100;
-        let object = basic_output(value, CoinType::Iota).unwrap();
+        let object = basic_output(value).unwrap();
         let gas_coin_balance = get_gas_balance_maybe(&object).unwrap();
         assert_eq!(gas_coin_balance.value(), value);
     }
