+++ crates/iota-types/src/stardust/error.rs
@@ -4,7 +4,7 @@
 //! Error types pertaining to deserializing Stardust snapshots
 use std::convert::Infallible;
 
-use iota_sdk::types::block::output::FoundryId;
+use iota_stardust_sdk::types::block::output::FoundryId;
 use packable::error::UnknownTagError;
 use thiserror::Error;
 
@@ -15,7 +15,7 @@ pub enum StardustError {
     #[error("invalid snapshot kind: {0}")]
     InvalidSnapshotKind(u8),
     #[error("block error: {0}")]
-    BlockError(#[from] iota_sdk::types::block::Error),
+    BlockError(#[from] iota_stardust_sdk::types::block::Error),
     #[error("{0}")]
     UnknownTag(#[from] UnknownTagError<u8>),
     #[error(
