+++ crates/iota-types/src/stardust/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod address;
+pub mod capped_coin;
+pub mod coin_kind;
+pub mod error;
+pub mod output;
+
+pub use address::*;
