+++ crates/iota-types/src/stardust/output/alias.rs
@@ -2,21 +2,20 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use iota_protocol_config::ProtocolConfig;
-use iota_sdk::types::block::output::AliasOutput as StardustAlias;
-use iota_types::{
+use iota_stardust_sdk::types::block::output::AliasOutput as StardustAlias;
+use move_core_types::{ident_str, identifier::IdentStr, language_storage::StructTag};
+use serde::{Deserialize, Serialize};
+use serde_with::serde_as;
+
+use crate::{
     balance::Balance,
     base_types::{IotaAddress, ObjectID, SequenceNumber, TxContext},
     collection_types::Bag,
     id::UID,
     object::{Data, MoveObject, Object, Owner},
+    stardust::stardust_to_iota_address,
     TypeTag, STARDUST_PACKAGE_ID,
 };
-use move_core_types::{ident_str, identifier::IdentStr, language_storage::StructTag};
-use serde::{Deserialize, Serialize};
-use serde_with::serde_as;
-
-use super::stardust_to_iota_address;
-use crate::stardust::migration::CoinType;
 
 pub const ALIAS_MODULE_NAME: &IdentStr = ident_str!("alias");
 pub const ALIAS_OUTPUT_MODULE_NAME: &IdentStr = ident_str!("alias_output");
@@ -145,7 +144,7 @@ pub struct AliasOutput {
     /// This is a "random" UID, not the AliasID from Stardust.
     pub id: UID,
 
-    /// The amount of coins held by the output.
+    /// The amount of IOTA coins held by the output.
     pub balance: Balance,
     /// The `Bag` holds native tokens, key-ed by the stringified type of the
     /// asset. Example: key: "0xabcded::soon::SOON", value:
@@ -185,14 +184,14 @@ impl AliasOutput {
         protocol_config: &ProtocolConfig,
         tx_context: &TxContext,
         version: SequenceNumber,
-        coin_type: &CoinType,
+        type_param: TypeTag,
     ) -> anyhow::Result<Object> {
         // Construct the Alias Output object.
         let move_alias_output_object = unsafe {
             // Safety: we know from the definition of `AliasOutput` in the stardust package
             // that it does not have public transfer (`store` ability is absent).
             MoveObject::new_from_execution(
-                AliasOutput::tag(coin_type.to_type_tag()).into(),
+                AliasOutput::tag(type_param).into(),
                 false,
                 version,
                 bcs::to_bytes(&self)?,
