+++ crates/iota-types/src/stardust/output/foundry.rs
@@ -2,33 +2,28 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use iota_protocol_config::ProtocolConfig;
-use iota_sdk::types::block::output::{FoundryOutput, OutputId};
-use iota_types::{
+use iota_stardust_sdk::types::block::output::{FoundryOutput, OutputId};
+
+use crate::{
     base_types::{ObjectID, SequenceNumber, TxContext},
     id::UID,
     object::Object,
+    stardust::address::stardust_to_iota_address,
 };
 
-use crate::stardust::{
-    migration::CoinType,
-    types::{output::create_coin, stardust_to_iota_address},
-};
-
-pub(crate) fn create_foundry_amount_coin(
+pub fn create_foundry_gas_coin(
     output_id: &OutputId,
     foundry: &FoundryOutput,
     tx_context: &TxContext,
     version: SequenceNumber,
     protocol_config: &ProtocolConfig,
-    coin_type: &CoinType,
 ) -> anyhow::Result<Object> {
-    create_coin(
+    crate::stardust::output::create_gas_coin(
         UID::new(ObjectID::new(output_id.hash())),
         stardust_to_iota_address(*foundry.alias_address())?,
         foundry.amount(),
         tx_context,
         version,
         protocol_config,
-        coin_type,
     )
 }
