+++ crates/iota-types/src/stardust/output/mod.rs
@@ -1,0 +1,8 @@
+pub mod alias;
+pub mod basic;
+pub mod foundry;
+pub mod nft;
+
+pub use alias::*;
+pub use basic::*;
+pub use nft::*;
